+++ src/components/Hero.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const Hero = (props) => (
+  <section className="hero is-warning is-fullheight is-bold">
+    <div className="hero-body">
+      <div className="container has-text-centered">
+        <h1 className="title has-text-centered">
+          CSS-in-JS
+        </h1>
+        <h2 className="subtitle">
+          Comparison of css-in-js solutions
+        </h2>
+      </div>
+    </div>
+  </section>
+);
+
+export default Hero;
